# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Elixir & Oak API",
    "version": "1.0.0",
    "description": "Complete REST API documentation for Elixir & Oak restaurant management platform. Covers all 17 route modules."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 base endpoint"
    }
  ],
  "paths": {
...
    "/contact": {
      "post": {
        "summary": "Submit contact message",
        "tags": [
          "Contact"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message submitted"
          }
        }
      }
    },
    "/newsletter": {
      "post": {
        "summary": "Subscribe to newsletter",
        "tags": [
          "Newsletter"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscribed"
          }
        }
      }
    },
    "/newsletter/unsubscribe": {
      "post": {
        "summary": "Unsubscribe",
        "tags": [
          "Newsletter"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsubscribed"
          }
        }
      }
    },
    "/admin/contact": {
      "get": {
        "summary": "List contact messages",
        "tags": [
          "Contact - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated messages"
          }
        }
      }
    },
    "/admin/contact/stats": {
      "get": {
        "summary": "Contact stats",
        "tags": [
          "Contact - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics"
          }
        }
      }
    },
    "/admin/contact/{id}": {
      "get": {
        "summary": "Get single message",
        "tags": [
          "Contact - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message details"
          }
        }
      },
      "delete": {
        "summary": "Delete message",
        "tags": [
          "Contact - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
    "/admin/contact/{id}/read": {
      "put": {
        "summary": "Mark message read/unread",
        "tags": [
          "Contact - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isRead": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        }
      }
    },
    "/admin/newsletter": {
      "get": {
        "summary": "List subscribers",
        "tags": [
          "Newsletter - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated subscribers"
          }
        }
      }
    },
    "/admin/newsletter/stats": {
      "get": {
        "summary": "Newsletter stats",
        "tags": [
          "Newsletter - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics"
          }
        }
      }
    },
    "/admin/newsletter/export": {
      "get": {
        "summary": "Export subscribers",
        "tags": [
          "Newsletter - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exported data"
          }
        }
      }
    },
    "/admin/newsletter/{id}": {
      "delete": {
        "summary": "Delete subscriber",
        "tags": [
          "Newsletter - Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
  }
}